Adopt unmigrate for schema reverts
